feat: simple std metric + comparisons property
